found some buggy behavior in the -1 input to ll insert and remove

+++ containers.h
@@ -64,7 +64,7 @@ typedef struct Con_Linked_List {
         
         Con_Linked_List_Node *head;
         Con_Linked_List_Node *sequential_access_ptr;
-        long long count;
+        int count;
         
         union Alloctaor alloctaor;
         union Dealloctaor dealloctaor;
